Move AbstractTaskViewModel to datacollection package
