+++ ground/src/main/java/com/google/android/ground/ui/datacollection/AbstractTaskViewModel.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.ui.editsubmission
+package com.google.android.ground.ui.datacollection
 
 import android.content.res.Resources
 import androidx.lifecycle.LiveData
