+++ ground/src/main/java/com/google/android/ground/ui/datacollection/TaskFragment.kt
@@ -15,8 +15,6 @@
  */
 package com.google.android.ground.ui.datacollection
 
-import com.google.android.ground.ui.editsubmission.AbstractTaskViewModel
-
 interface TaskFragment<T : AbstractTaskViewModel> {
   var viewModel: T
