+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/photo/PhotoTaskViewModel.kt
@@ -25,7 +25,7 @@ import com.google.android.ground.model.task.Task
 import com.google.android.ground.persistence.remote.firebase.FirebaseStorageManager.Companion.getRemoteMediaPath
 import com.google.android.ground.repository.UserMediaRepository
 import com.google.android.ground.rx.annotations.Hot
-import com.google.android.ground.ui.editsubmission.AbstractTaskViewModel
+import com.google.android.ground.ui.datacollection.AbstractTaskViewModel
 import com.google.android.ground.ui.editsubmission.PhotoResult
 import com.google.android.ground.ui.util.BitmapUtil
 import io.reactivex.Observable
